requirements.txt
fastapi
uvicorn[standard]

pydantic
python-multipart

sqlalchemy
psycopg2-binary
alembic

passlib[bcrypt]
python-jose[cryptography]

torch
torchvision
numpy
pillow

transformers
sentence-transformers
scikit-learn

langchain
faiss-cpu

requests
httpx

pandas

python-dotenv

aiofiles

matplotlib

prometheus-fastapi-instrumentator